Internal branch sync. Included in this change:
. lib/file - add File_MakeSafeTempDir()
. changes in shared code that don't affect open-vm-tools functionality

+++ open-vm-tools/modules/shared/vmxnet/eth_public.h
@@ -1143,7 +1143,7 @@ Eth_GetPayload(const void *frame)
 
 static INLINE Bool
 Eth_IsFrameHeaderComplete(const Eth_Header *eh,
-                          const uint16 len,
+                          const uint32 len,
                           uint16 *ehHdrLen)
 {
    uint16 ehLen;
